% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.leedid}
\alias{print.leedid}
\title{Print a \code{leedid} object}
\usage{
\method{print}{leedid}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{leedid} object, as returned by \code{\link[=leedid]{leedid()}}.}

\item{digits}{Number of digits to round to.}

\item{...}{Unused, present for S3 consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{leedid} object
}
